fix phpstan errors

+++ backend/database/seeders/DatabaseSeeder.php
@@ -31,16 +31,16 @@ class DatabaseSeeder extends Seeder
         $mapVertices = Map::vertices();
         /** @var Vertex $vertex */
         foreach ($mapVertices as $vertex) {
-            $id = (int)Str::remove(search: 'n', subject: (string)$vertex->getId());
+            $id = Str::remove(search: 'n', subject: (string)$vertex->getId());
             if (MapVertex::find($id) === null) {
                 MapVertex::factory(1)->create(['id' => $id]);
             }
         }
 
         if (TransportRequest::all()->count() === 0) {
-            /** @var User $user */
             $userBar = $output->createProgressBar(count(User::all()));
             $userBar->start();
+            /** @var User $user */
             foreach (User::all() as $user) {
                 if ($user->isAuctioneer()) {
                     continue;
